Move /build tests functions from main_test to view_build_test.py
Also, delete some pr tests that were previously moved.
