+++ gubernator/main_test.py
@@ -21,7 +21,6 @@ To run these tests:
     $ nosetests --with-gae --gae-lib-root ~/google_appengine/
 """
 
-import os
 import unittest
 
 import webtest
@@ -31,7 +30,6 @@ import cloudstorage as gcs
 import main
 import gcs_async
 import gcs_async_test
-import view_pr
 
 write = gcs_async_test.write
 
@@ -81,114 +79,10 @@ class AppTest(TestBase):
         self.init_stubs()
         init_build(self.BUILD_DIR)
 
-    def get_build_page(self):
-        return app.get('/build' + self.BUILD_DIR)
-
     def test_index(self):
         """Test that the index works."""
         response = app.get('/')
         self.assertIn('kubernetes-e2e-gce', response)
-
-    def test_missing(self):
-        """Test that a missing build gives a 404."""
-        response = app.get('/build' + self.BUILD_DIR.replace('1234', '1235'),
-                           status=404)
-        self.assertIn('1235', response)
-
-    def test_missing_started(self):
-        """Test that a missing started.json still renders a proper page."""
-        build_dir = '/kubernetes-jenkins/logs/job-with-no-started/1234/'
-        init_build(build_dir, started=False)
-        response = app.get('/build' + build_dir)
-        self.assertIn('Build Result: SUCCESS', response)
-        self.assertIn('job-with-no-started', response)
-        self.assertNotIn('Started', response)  # no start timestamp
-        self.assertNotIn('github.com', response)  # no version => no src links
-
-    def test_missing_finished(self):
-        """Test that a missing finished.json still renders a proper page."""
-        build_dir = '/kubernetes-jenkins/logs/job-still-running/1234/'
-        init_build(build_dir, finished=False)
-        response = app.get('/build' + build_dir)
-        self.assertIn('Build Result: Not Finished', response)
-        self.assertIn('job-still-running', response)
-        self.assertIn('Started', response)
-
-    def test_build(self):
-        """Test that the build page works in the happy case."""
-        response = self.get_build_page()
-        self.assertIn('2014-07-28', response)  # started
-        self.assertIn('16m40s', response)      # build duration
-        self.assertIn('Third', response)       # test name
-        self.assertIn('1m36s', response)       # test duration
-        self.assertIn('Build Result: SUCCESS', response)
-        self.assertIn('Error Goes Here', response)
-        self.assertIn('test.go#L123">', response)  # stacktrace link works
-
-    def test_build_no_failures(self):
-        """Test that builds with no Junit artifacts work."""
-        gcs.delete(self.BUILD_DIR + 'artifacts/junit_01.xml')
-        response = self.get_build_page()
-        self.assertIn('No Test Failures', response)
-
-    def test_build_show_log(self):
-        """Test that builds that failed with no failures show the build log."""
-        gcs.delete(self.BUILD_DIR + 'artifacts/junit_01.xml')
-        write(self.BUILD_DIR + 'finished.json',
-              {'result': 'FAILURE', 'timestamp': 1406536800})
-
-        # Unable to fetch build-log.txt, still works.
-        response = self.get_build_page()
-        self.assertNotIn('Error lines', response)
-
-        self.testbed.init_memcache_stub()  # clear cached result
-        write(self.BUILD_DIR + 'build-log.txt',
-              u'ERROR: test \u039A\n\n\n\n\n\n\n\n\nblah'.encode('utf8'))
-        response = self.get_build_page()
-        self.assertIn('Error lines', response)
-        self.assertIn('No Test Failures', response)
-        self.assertIn('ERROR</span>: test', response)
-        self.assertNotIn('blah', response)
-
-    def test_build_failure_no_text(self):
-        # Some failures don't have any associated text.
-        write(self.BUILD_DIR + 'artifacts/junit_01.xml', '''
-            <testsuites>
-                <testsuite tests="1" failures="1" time="3.274" name="k8s.io/test/integration">
-                    <testcase classname="integration" name="TestUnschedulableNodes" time="0.210">
-                        <failure message="Failed" type=""/>
-                    </testcase>
-                </testsuite>
-            </testsuites>''')
-        response = self.get_build_page()
-        self.assertIn('TestUnschedulableNodes', response)
-        self.assertIn('junit_01.xml', response)
-
-    def test_build_pr_link(self):
-        ''' The build page for a PR build links to the PR results.'''
-        build_dir = '/%s/123/e2e/567/' % view_pr.PR_PREFIX
-        init_build(build_dir)
-        response = app.get('/build' + build_dir)
-        self.assertIn('PR #123', response)
-        self.assertIn('href="/pr/123"', response)
-
-    def test_cache(self):
-        """Test that caching works at some level."""
-        response = self.get_build_page()
-        gcs.delete(self.BUILD_DIR + 'started.json')
-        gcs.delete(self.BUILD_DIR + 'finished.json')
-        response2 = self.get_build_page()
-        self.assertEqual(str(response), str(response2))
-
-    def test_build_list(self):
-        """Test that the job page shows a list of builds."""
-        response = app.get('/builds' + os.path.dirname(self.BUILD_DIR[:-1]))
-        self.assertIn('/1234/">1234</a>', response)
-
-    def test_job_list(self):
-        """Test that the job list shows our job."""
-        response = app.get('/jobs/kubernetes-jenkins/logs')
-        self.assertIn('somejob/">somejob</a>', response)
 
     def test_nodelog_missing_files(self):
         """Test that a missing all files gives a 404."""
@@ -309,57 +203,3 @@ class AppTest(TestBase):
                     '01-01T01:01:01.005Z last line')
         response = app.get('/build' + nodelog_url)
         self.assertIn(expected, response)
-
-
-class PRTest(TestBase):
-    BUILDS = {
-        'build': [('12', {'version': 'bb', 'timestamp': 1467147654}, None),
-                  ('11', {'version': 'bb', 'timestamp': 1467146654}, {'result': 'PASSED'}),
-                  ('10', {'version': 'aa', 'timestamp': 1467136654}, {'result': 'FAILED'})],
-        'e2e': [('47', {'version': 'bb', 'timestamp': '1467147654'}, {'result': '[UNSET]'}),
-                ('46', {'version': 'aa', 'timestamp': '1467136700'}, {'result': '[UNSET]'})]
-    }
-
-    def setUp(self):
-        self.init_stubs()
-
-    def init_pr_directory(self):
-        gcs_async_test.install_handler(self.testbed.get_stub('urlfetch'),
-            {'123/': ['build', 'e2e'],
-             '123/build/': ['11', '10', '12'],  # out of order
-             '123/e2e/': ['47', '46']})
-
-        for job, builds in self.BUILDS.iteritems():
-            for build, started, finished in builds:
-                path = '/%s/123/%s/%s/' % (view_pr.PR_PREFIX, job, build)
-                if started:
-                    write(path + 'started.json', started)
-                if finished:
-                    write(path + 'finished.json', finished)
-
-    def test_pr_builds(self):
-        self.init_pr_directory()
-        builds = view_pr.pr_builds('123')
-        self.assertEqual(builds, self.BUILDS)
-
-    def test_pr_handler(self):
-        self.init_pr_directory()
-        response = app.get('/pr/123')
-        self.assertIn('e2e/47', response)
-        self.assertIn('PASSED', response)
-        self.assertIn('colspan="3"', response)  # header
-        self.assertIn('github.com/kubernetes/kubernetes/pull/123', response)
-        self.assertIn('28 20:44', response)
-
-    def test_pr_handler_missing(self):
-        gcs_async_test.install_handler(self.testbed.get_stub('urlfetch'),
-            {'124/': []})
-        response = app.get('/pr/124')
-        self.assertIn('No Results', response)
-
-    def test_pr_build_log_redirect(self):
-        path = '123/some-job/55/build-log.txt'
-        response = app.get('/pr/' + path)
-        self.assertEqual(response.status_code, 302)
-        self.assertIn('https://storage.googleapis.com', response.location)
-        self.assertIn(path, response.location)
